cpufreq: imx6q: read OCOTP through nvmem in fallback case
The imx-ocotp nvmem driver now resets the clock during init, so
direct accesses that used to work now block because of missing
clock.
Always go through the nvmem driver, but look for the driver
directly in case the nvmem cells are not defined in dtb.

If that also fails, just skip setting cpufreq max speed with
a warning: that is better than not booting at all.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>